collisions using rect
currently not bouncing correctly. would like to have the player be able to control the bounce a little more. if it hits top half will go up or bottom half will go down. at the moment both hitboxes are complete but there is a gap in the middle.  also added a bounce variable in an attempt to remove the multibounce possibility.